Disable application heartbeats
During a graceful (warm) shutdown, tasks are allowed to complete before
the worker terminates. This allows in-progress tasks to complete.

When acks_late is enabled, those tasks should ACK the message to RabbitMQ
so they aren't redelivered to another worker.

But, celery won't send application heartbeats during a warm shutdown. So,
if you have a long running task running during a shutdown, RabbitMQ will
detect the lack of heartbeat, shutdown the connection, and redeliver
the message to a new worker.

Disabling application heartbeats allows the connection to stay open, and
successfully send the ACK back to RabbitMQ so that it can remove the
message from the queue.